+++ scripts/ageing_society/au_exp_survey_clean.py
@@ -22,7 +22,7 @@ load_dotenv()
 
 # Read the variable
 data_home = Path(os.getenv("DATA_HOME"))
-current_version = os.getenv("CURRENT_VERSION")
+current_version = os.getenv(f"CURRENT_VERSION_AGEING_SOCIETY")
 
 timestamp = datetime.datetime.now()
 file_timestamp = timestamp.ctime()
@@ -43,12 +43,19 @@ current_module = "ageing_society"
 
 logging.info("Configure data source")
 data_source = "au_exp_survey"
+country = "australia"
 
 logging.info("Configure paths")
 path_data_raw = data_home / "raw_data" / current_module / current_version / data_source
 path_data_clean = (
     data_home / "clean_data" / current_module / current_version / data_source
 )
+
+logging.info("Load the currency convert rates")
+file_name_conv_rate = [file_ for file_ in path_data_raw.parent.glob("currency*")][0]
+raw_conv_rate = pd.read_csv(file_name_conv_rate)
+raw_conv_rate_group = raw_conv_rate.groupby("time")
+
 
 if not path_data_clean.exists():
     path_data_clean.mkdir(parents=True, exist_ok=True)
@@ -171,9 +178,17 @@ for data_path in path_data_raw.glob("*.xls*"):
             exp_age_cohort = raw_exp_unique.iloc[1:, [0, pos]]
             househouse_size = raw_exp_unique.iloc[0, pos]
 
+            raw_conv_rate_ = raw_conv_rate_group.get_group(int(year))[
+                f"{country}, convert rate local currency to 2005 usd"
+            ].values[0]
             exp_age_cohort["value"] = (
-                np.array(exp_age_cohort[column]) / househouse_size * 365 / 7
-            )  # convert to annual expenditure
+                np.array(exp_age_cohort[column])
+                / househouse_size
+                * 365
+                / 7
+                * raw_conv_rate_
+            )  # convert to annual expenditure, and to 2005 usd
+
             exp_age_cohort["value"] = pd.to_numeric(
                 exp_age_cohort["value"], errors="coerce"
             )
@@ -183,9 +198,9 @@ for data_path in path_data_raw.glob("*.xls*"):
             cleaned_exp_per_capita = pd.concat(
                 [cleaned_exp_per_capita, exp_age_cohort.drop(columns=[column])]
             )
-            del exp_age_cohort, househouse_size, pos
+            del exp_age_cohort, househouse_size, pos, raw_conv_rate_
     cleaned_exp_per_capita["time"] = year
-    cleaned_exp_per_capita["unit"] = "current prices, australian dollars"
+    cleaned_exp_per_capita["unit"] = "2005 usd"
 
     logging.info("Concat cleaned data from different years")
     cleaned_exp_full = pd.concat([cleaned_exp_full, cleaned_exp_per_capita])
@@ -194,6 +209,6 @@ for data_path in path_data_raw.glob("*.xls*"):
 
 # save data
 logging.info("Save cleaned data")
-cleaned_exp_full["country"] = "australia"
-file_name = f"{data_source}_household_data_all.csv"
+cleaned_exp_full["country"] = country
+file_name = f"{data_source}_household_data_all_in_2005_usd.csv"
 cleaned_exp_full.to_csv(path_data_clean / file_name, index=False)